libdvdcss: Avoid a cast when comparing the return values of write() and strlen()

+++ src/libdvdcss.c
@@ -375,8 +375,8 @@ LIBDVDCSS_EXPORT dvdcss_t dvdcss_open ( char *psz_target )
         i_fd = open( psz_tagfile, O_RDWR|O_CREAT, 0644 );
         if( i_fd >= 0 )
         {
-            size_t len = strlen(psz_tag);
-            if( write( i_fd, psz_tag, len ) < (long)len )
+            ssize_t len = strlen(psz_tag);
+            if( write( i_fd, psz_tag, len ) < len )
             {
                 print_error( dvdcss,
                              "Error writing cache directory tag, continuing..\n" );
